feat: move create market setup to libs/cypress
