+++ apps/trading-e2e/cypress.config.js
@@ -24,14 +24,20 @@ module.exports = defineConfig({
     requestTimeout: 20000,
   },
   env: {
-    ETHEREUM_PROVIDER_URL:
-      'https://sepolia.infura.io/v3/4f846e79e13f44d1b51bbd7ed9edefb8',
+    // pubkey and token for capsule_wallet
+    CAPSULE_VEGA_PUBLIC_KEY:
+      '02eceaba4df2bef76ea10caf728d8a099a2aa846cced25737cccaa9812342f65',
+    CAPSULE_VEGA_WALLET_API_TOKEN:
+      '1MRzTpKL7vDgHctjUN3bwWYsOUcLwTbKzTwVW7oEGcALd4s2KqU1RDgGlPg9GhnB',
+    // pubkeys for UI_Trading_Test
     VEGA_PUBLIC_KEY:
       '47836c253520d2661bf5bed6339c0de08fd02cf5d4db0efee3b4373f20c7d278',
     VEGA_PUBLIC_KEY2:
       '1a18cdcaaa4f44a57b35a4e9b77e0701c17a476f2b407620f8c17371740cf2e4',
     TRUNCATED_VEGA_PUBLIC_KEY: '47836c…c7d278',
     TRUNCATED_VEGA_PUBLIC_KEY2: '1a18cd…0cf2e4',
+    ETHEREUM_PROVIDER_URL:
+      'https://sepolia.infura.io/v3/4f846e79e13f44d1b51bbd7ed9edefb8',
     ETHEREUM_WALLET_ADDRESS: '0x265Cc6d39a1B53d0d92068443009eE7410807158',
     ETHERSCAN_URL: 'https://sepolia.etherscan.io',
     ETHEREUM_CHAIN_ID: 11155111,
