+++ libs/cypress/src/index.ts
@@ -15,6 +15,8 @@ import {
   addSetVegaWallet,
 } from './lib/commands/vega-wallet-connect';
 import { addMockTransactionResponse } from './lib/commands/mock-transaction-response';
+import { addCreateMarket } from './lib/commands/create-market';
+import { addSetVegaWalletConfig } from './lib/commands/set-vega-wallet-config';
 
 addGetTestIdcommand();
 addSlackCommand();
@@ -31,11 +33,14 @@ addUpdateCapsuleMultiSig();
 addVegaWalletConnect();
 addSetVegaWallet();
 addMockTransactionResponse();
+addCreateMarket();
+addSetVegaWalletConfig();
 
 export { mockConnectWallet } from './lib/commands/vega-wallet-connect';
 export type { onMessage } from './lib/mock-ws';
 export { aliasGQLQuery } from './lib/mock-gql';
 export { aliasWalletQuery } from './lib/mock-rest';
+export * from './lib/utils';
 
 Cypress.on(
   'uncaught:exception',
