+++ libs/cypress/src/lib/capsule/faucet-asset.ts
